requirements.txt
alembic==0.9.10
Flask==1.0
flask-migrate==2.2.1
flask_restful==0.3.6
flask-script==2.0.6
flask-sqlalchemy==2.3.2
Mako==1.0.7
pep8==1.7.1
pymemcache==2.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytest==3.8.2
pytest-mock==1.10.0
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
request==1.0.2
six==1.11.0
webargs==4.1.0
